requirements.txt
Django==4.2.10
djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.5
pygments==2.17.2
pytest==7.4.4
pytest-django==4.7.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
pyyaml==6.0.1
uritemplate==4.1.1 